Enhancements to node label look-and-feel (#50)
There was an issue with svg text-rendering for system-ui fonts,
so we used the default monospace as an alternative.

This fixes the text rendering issue, so we can switch back to the sans font.
I've also shrunk the node icon to make it less huge. With the combination
of the smaller icon and a non-monospaced font, we now have much more
room for the label.