Add OneOfControl for properties with multiple types
Adds a new OneOfControl that allows a property to have multiple types
and let a user choice which type they want to use then display the correct
form field.

This allows for support of properties that can either be entered directly or
taken from a parent node's output.

Paired with https://github.com/elyra-ai/elyra/pull/2244