Add support for opening component definition
Adds a new case for the `openFile` action that will instead trigger
the `openComponentDef` when a node has no defined `filenameRef`
(ie node's without a related file).

Handles the edge case when a `filenameRef` exists but the the
referenced property is not set the "Open File" option exists but
is disabled (as it was previously).

Elyra PR: https://github.com/elyra-ai/elyra/pull/2564